refactor(hooks): move hooks to root directory
