+++ frontend/src/app/(main)/chat/Home.tsx
@@ -17,9 +17,9 @@ import { GET_CHAT_HISTORY } from '@/graphql/request';
 import { useMutation, useQuery } from '@apollo/client';
 import { toast } from 'sonner';
 import { EventEnum } from '@/components/enum';
-import { useModels } from '../../hooks/useModels';
-import { useChatList } from '../../hooks/useChatList';
-import { useChatStream } from '../../hooks/useChatStream';
+import { useModels } from '../../../hooks/useModels';
+import { useChatList } from '../../../hooks/useChatList';
+import { useChatStream } from '../../../hooks/useChatStream';
 import EditUsernameForm from '@/components/edit-username-form';
 import ChatContent from '@/components/chat/chat';
 import { ProjectContext } from '@/components/code-engine/project-context';
