+++ frontend/src/app/(main)/chat/MainLayout.tsx
@@ -1,6 +1,5 @@
 'use client';
 import React, { useEffect, useState } from 'react';
-import { useAuthContext } from '@/app/providers/AuthProvider';
 import { useRouter } from 'next/navigation';
 import { ResizablePanel, ResizablePanelGroup } from '@/components/ui/resizable';
 import { SidebarProvider } from '@/components/ui/sidebar';
@@ -9,8 +8,9 @@ import { ChatSideBar } from '@/components/sidebar';
 import ProjectModal from '@/components/project-modal';
 import { useQuery } from '@apollo/client';
 import { ProjectProvider } from '@/components/code-engine/project-context';
-import { useChatList } from '@/app/hooks/useChatList';
+import { useChatList } from '@/hooks/useChatList';
 import { GET_USER_PROJECTS } from '@/graphql/request';
+import { useAuthContext } from '@/providers/AuthProvider';
 
 export default function MainLayout({
   children,
