+++ frontend/src/hooks/useSpeechRecognition.ts
